Fix cli:seed to translate scenario data to the load endpoint's wire format
The single-object shape migration left cli/seed.cli.js posting scenario data straight to the seed endpoint, but the endpoint still expects pluralized, array-valued keys. The Playwright load fixture already had a private translator for this in tests/support/fixtures.js, so it's extracted to a shared tests/support/helpers/wire-format.helpers.js and used by both the fixture and the CLI.
